refactor: code cleanup and refactoring
